chore(android): Code cleanup
